Refactor: BrowseMainViewController의 browseMainCollectionview BrowseMainCollectionView로 분리
